requirements.txt
fastapi
slowapi
uvicorn
aiofiles
python-dotenv